[design] 나의 대표 고양이 UI 개선
